extracted writing views and composer
