+++ frontend/src/main.jsx
@@ -1,16 +1,5 @@
-import {
-  lazy,
-  useState,
-  useRef,
-  useEffect,
-  useMemo,
-  useCallback,
-  memo,
-  useId,
-  useLayoutEffect,
-  Suspense,
-} from "react";
-import { cx, CONTROL_MOTION, FADE_MOTION, PROMPT_BAR_CONTROL_MOTION } from "./uiShared.js";
+import { cx, CONTROL_MOTION, FADE_MOTION } from "./uiShared.js";
+import { useState, useRef, useEffect, useMemo, useCallback, memo } from "react";
 import {
   formatInteger,
   formatCost,
@@ -31,10 +20,6 @@ import {
   ChevronDown,
   Copy,
   RefreshCw,
-  ArrowLeft,
-  Square,
-  Plus,
-  Check,
   Menu,
 } from "lucide-react";
 import { MaskIcon, IconButton } from "./components/IconButton.jsx";
@@ -58,32 +43,6 @@ import { uncitedSources } from "./websearch/citations.js";
 import { useStreamReveal } from "./streamingText.js";
 import AttachmentChips from "./attachments/AttachmentChips.jsx";
 import SourcePills from "./websearch/SourcePills.jsx";
-import { useRafScroller } from "./streamScroll.js";
-import { ThinkingStatus } from "./thinkingStates.jsx";
-import StoryLorebook from "./lorebook/StoryLorebook.jsx";
-import ChapterStreamingCanvas from "./writing/ChapterStreamingCanvas.jsx";
-import {
-  supportsThinking,
-  supportsImageInput,
-  requiresThinking,
-  effectiveThinkingEnabled,
-  reasoningEffortLabel,
-} from "./modelReasoning.js";
-import AttachButton from "./attachments/AttachButton.jsx";
-import { MAX_FILES_PER_MESSAGE } from "./attachments/attachmentsApi.js";
-import { ContextWindowMeter } from "./components/ContextWindowMeter.jsx";
-import { SystemPromptModal } from "./settings/SystemPromptModal.jsx";
-import {
-  historyRunGroups,
-  isPromptEntry,
-  historyCostTotal,
-  historyWordTotals,
-  historyRunRows,
-  historyRowChildren,
-  historyRowText,
-  historyRowWords,
-  historyRowLabel,
-} from "./writing/historyFormatting.js";
 import {
   readLocalAppSettings,
   pickOpeningMessage,
@@ -97,9 +56,16 @@ import { useTour } from "./tour/useTour.js";
 import { WRITE_TOUR_STEPS } from "./tour/tourSteps.js";
 import { useNotifications } from "./notifications/useNotifications.js";
 import { useAttachments } from "./attachments/useAttachments.js";
+import {
+  supportsImageInput,
+  requiresThinking,
+  effectiveThinkingEnabled,
+  reasoningEffortLabel,
+} from "./modelReasoning.js";
 import { createNavigationCoordinator } from "./writing/navigationCoordinator.js";
 import { createSaveCoordinator } from "./writing/saveCoordinator.js";
 import { storyApi } from "./writing/storyApi.js";
+import { useRafScroller } from "./streamScroll.js";
 import {
   chapterRunTargetsOpenChapter,
   loadSettledGeneration,
@@ -119,29 +85,23 @@ import { generateLorebookEntry as generateLorebookEntryStream } from "./lorebook
 import { useFileDrop } from "./attachments/useFileDrop.js";
 import HelpTourButton from "./HelpTourButton.jsx";
 import StoryBrainstorm from "./brainstorm/StoryBrainstorm.jsx";
+import { ContextWindowMeter } from "./components/ContextWindowMeter.jsx";
+import { StoryWorkspace } from "./writing/StoryWorkspace.jsx";
+import { LOREBOOK_PHASE_LABELS } from "./writing/WriteOperationStatus.jsx";
+import { WriteLanding } from "./writing/WriteLanding.jsx";
+import { Composer } from "./composer/Composer.jsx";
 import { SettingsDrawer } from "./settings/SettingsDrawer.jsx";
 import { ConfirmModal } from "./components/ConfirmModal.jsx";
+import { NewStoryModal } from "./writing/NewStoryModal.jsx";
 import NotificationStack from "./notifications/NotificationStack.jsx";
 import TourOverlay from "./tour/TourOverlay.jsx";
 import { TosLoadingScreen, TosUnavailableScreen, TosGateModal } from "./TosGate.jsx";
 import { createRoot } from "react-dom/client";
 import "./styles.css";
 
-const ChapterCanvasEditor = lazy(() => import("./writing/ChapterCanvasEditor.jsx"));
-
 const FEEDBACK_FORM_URL = "https://forms.gle/gTth2TcXLYAArvGm6";
 
 const GITHUB_RELEASES_LATEST_URL = "https://api.github.com/repos/echo1097/routerchat/releases/latest";
-
-const WRITE_GENERATION_MODES = {
-  edit: "Edit Chapter",
-  new: "New Chapter",
-};
-
-const LOREBOOK_UPDATE_MODES = {
-  auto: "Auto Lorebook",
-  manual: "Manual Lorebook",
-};
 
 function AssistantActionButton({ label, children, ...props }) {
   return (
@@ -2052,20 +2012,6 @@ function EmptyChatState() {
   );
 }
 
-function WriteLanding({ openingMessage }) {
-  if (!openingMessage) return null;
-
-  return (
-    <section className="pointer-events-none absolute inset-0 z-10 flex items-center justify-center px-4 pb-[12vh] pt-20 sm:px-8 lg:px-10">
-      <div className="mx-auto flex w-full max-w-[760px] flex-col items-center">
-        <div className="text-balance text-center text-[28px] font-medium leading-tight tracking-[-0.02em] text-neutral-200 sm:text-[42px]">
-          {openingMessage}
-        </div>
-      </div>
-    </section>
-  );
-}
-
 function TemporaryChatButton({ active, onClick }) {
   return (
     <button
@@ -2654,1392 +2600,6 @@ function StoryRail({
         onClose={() => setSearchOpen(false)}
       />
     </>
-  );
-}
-
-function editPreviewPhrase(operation, anchor) {
-  const snippet = anchor ? `“${anchor.length > 60 ? `${anchor.slice(0, 60)}…` : anchor}”` : "";
-  switch (operation) {
-    case "replaceBlock":
-      return snippet ? `Rewriting near ${snippet}` : "Rewriting a paragraph";
-    case "insertBeforeBlock":
-      return snippet ? `Inserting before ${snippet}` : "Inserting a paragraph";
-    case "insertAfterBlock":
-      return snippet ? `Inserting after ${snippet}` : "Inserting a paragraph";
-    case "replaceBlockRange":
-      return "Replacing a range";
-    case "appendToChapter":
-      return "Appending to the chapter";
-    default:
-      return "Writing an edit";
-  }
-}
-
-const WRITE_STATUS_STATES = [
-  "Preparing",
-  "Working",
-  "Thinking",
-  "Writing",
-  "Fixing the edit",
-  "Updating Lorebook",
-  "Reconciling",
-];
-
-const LOREBOOK_PHASE_LABELS = {
-  working: "Working",
-  thinking: "Thinking",
-  updating: "Updating Lorebook",
-};
-
-function WriteOperationStatus({
-  status,
-  reasoning = "",
-  reasoningStreaming = false,
-  reasoningDurationMs = null,
-  editPreview = null,
-}) {
-  const [detailsOpen, setDetailsOpen] = useState(false);
-  const dismissedRef = useRef(false); //once you close it yourself we stop reopening it on you
-  const toggleRef = useRef(null);
-  const popoverRef = useRef(null);
-  const reasoningScrollRef = useRef(null);
-  const editScrollRef = useRef(null);
-  const [popoverPosition, setPopoverPosition] = useState(null);
-  const popoverId = useId();
-  const {
-    markUserScroll: markReasoningScroll,
-    markWheelIntent: markReasoningWheelIntent,
-    markTouchStart: markReasoningTouchStart,
-    markTouchMove: markReasoningTouchMove,
-    scrollToBottom: scrollReasoningToBottom,
-    startFollowing: startReasoningFollowing,
-  } = useRafScroller(reasoningScrollRef, 32);
-  const {
-    markUserScroll: markEditScroll,
-    markWheelIntent: markEditWheelIntent,
-    markTouchStart: markEditTouchStart,
-    markTouchMove: markEditTouchMove,
-    scrollToBottom: scrollEditToBottom,
-    startFollowing: startEditFollowing,
-  } = useRafScroller(editScrollRef, 32);
-
-  const hasReasoning = Boolean(reasoning);
-  const reasoningLabel =
-    !reasoningStreaming && reasoningDurationMs
-      ? `Thought for ${formatThoughtDuration(reasoningDurationMs)}`
-      : "Thinking";
-
-  //cannot gate on operation alone, replaceBlockRange puts it last in the schema so its prose arrives first
-  const hasEditPreview = Boolean(editPreview && (editPreview.newText || editPreview.operation));
-  const editLabel = editPreview ? `Edit ${editPreview.editIndex + 1}` : "";
-  const editPhrase = editPreview ? editPreviewPhrase(editPreview.operation, editPreview.anchor) : "";
-  const editStreaming = Boolean(editPreview && !editPreview.newTextComplete);
-
-  const hasDetails = hasReasoning || hasEditPreview;
-
-  const positionPopover = useCallback(() => {
-    const toggle = toggleRef.current;
-    if (!detailsOpen || !toggle) return;
-
-    const margin = 16;
-    const gap = 8;
-    const maxWidth = 480;
-    const rect = toggle.getBoundingClientRect();
-    const width = Math.min(maxWidth, window.innerWidth - margin * 2);
-    const left = Math.min(
-      Math.max(rect.right - width, margin),
-      window.innerWidth - width - margin,
-    );
-    const top = Math.max(rect.bottom + gap, margin);
-
-    setPopoverPosition({
-      top,
-      left,
-      width,
-      maxHeight: Math.max(160, window.innerHeight - top - margin),
-    });
-  }, [detailsOpen]);
-
-  //the whole point of the edit preview is that you can watch it, so open the panel yourself the first time
-  //prose shows up, otherwise an edit run looks exactly like the old mute status label it was meant to replace
-  useEffect(() => {
-    if (!hasEditPreview || dismissedRef.current) return;
-    setDetailsOpen(true);
-  }, [hasEditPreview]);
-
-  useEffect(() => {
-    if (!detailsOpen) return;
-    startReasoningFollowing();
-    startEditFollowing();
-  }, [detailsOpen, startReasoningFollowing, startEditFollowing]);
-
-  useLayoutEffect(() => {
-    if (!detailsOpen) {
-      setPopoverPosition(null);
-      return undefined;
-    }
-
-    positionPopover();
-    function repositionForScroll(event) {
-      if (popoverRef.current?.contains(event.target)) return;
-      positionPopover();
-    }
-    window.addEventListener("resize", positionPopover);
-    window.addEventListener("scroll", repositionForScroll, true);
-
-    return () => {
-      window.removeEventListener("resize", positionPopover);
-      window.removeEventListener("scroll", repositionForScroll, true);
-    };
-  }, [detailsOpen, positionPopover]);
-
-  useEffect(() => {
-    if (!detailsOpen) return;
-    scrollReasoningToBottom();
-  }, [reasoning, detailsOpen, scrollReasoningToBottom]);
-
-  useEffect(() => {
-    if (!detailsOpen) return;
-    scrollEditToBottom();
-  }, [editPreview, detailsOpen, scrollEditToBottom]);
-
-  const detailsPopover = hasDetails && detailsOpen && popoverPosition
-    ? createPortal(
-      <span
-        id={popoverId}
-        ref={popoverRef}
-        role="region"
-        aria-label="Writing details"
-        className="t-dropdown write-thinking-popover is-open"
-        data-origin="top-right"
-        style={popoverPosition}
-      >
-        {hasReasoning && (
-          <>
-            <span className="write-thinking-popover-header">
-              <span className="write-thinking-popover-heading">
-                <span
-                  className={cx("write-thinking-popover-title", reasoningStreaming && "t-shimmer")}
-                  data-text={reasoningLabel}
-                >
-                  {reasoningLabel}
-                </span>
-              </span>
-            </span>
-            <span
-              ref={reasoningScrollRef}
-              onScroll={markReasoningScroll}
-              onWheel={markReasoningWheelIntent}
-              onTouchStart={markReasoningTouchStart}
-              onTouchMove={markReasoningTouchMove}
-              className="write-thinking-popover-content"
-              data-testid="write-thinking-scroll"
-            >
-              <ThinkingContent>{reasoning}</ThinkingContent>
-            </span>
-          </>
-        )}
-        {hasReasoning && hasEditPreview && <span className="write-thinking-popover-divider" />}
-        {hasEditPreview && (
-          <>
-            <span className="write-thinking-popover-header">
-              <span className="write-thinking-popover-heading">
-                <span
-                  className={cx("write-thinking-popover-title", editStreaming && "t-shimmer")}
-                  data-text={editLabel}
-                >
-                  {editLabel}
-                </span>
-                <span className="write-thinking-popover-subtitle">{editPhrase}</span>
-              </span>
-            </span>
-            <span
-              ref={editScrollRef}
-              onScroll={markEditScroll}
-              onWheel={markEditWheelIntent}
-              onTouchStart={markEditTouchStart}
-              onTouchMove={markEditTouchMove}
-              className="write-thinking-popover-content"
-              data-testid="write-edit-preview-scroll"
-            >
-              <ThinkingContent>{editPreview.newText}</ThinkingContent>
-            </span>
-          </>
-        )}
-      </span>,
-      document.body,
-    )
-    : null;
-
-  return (
-    <>
-      <span className="write-operation-wrap">
-        <span className="write-operation-status" aria-live="polite">
-          <button
-            ref={toggleRef}
-            type="button"
-            onClick={hasDetails ? () => setDetailsOpen((value) => {
-              if (value) dismissedRef.current = true;
-              return !value;
-            }) : undefined}
-            disabled={!hasDetails}
-            className={cx(
-              "write-operation-thinking-toggle",
-              !hasDetails && "is-static",
-              CONTROL_MOTION,
-            )}
-            aria-label={detailsOpen ? "Collapse writing details" : "Expand writing details"}
-            aria-expanded={hasDetails ? detailsOpen : undefined}
-            aria-controls={hasDetails ? popoverId : undefined}
-          >
-            <ThinkingStatus
-              label={status}
-              states={WRITE_STATUS_STATES}
-              align="right"
-              className="write-operation-label"
-            />
-            <ChevronDown
-              size={14}
-              aria-hidden="true"
-              className={cx(
-                "write-operation-thinking-chevron transition-[transform,opacity] duration-[var(--dropdown-open-dur)] ease-[var(--dropdown-ease)]",
-                !detailsOpen && "-rotate-90",
-                !hasDetails && "opacity-0",
-              )}
-            />
-          </button>
-        </span>
-      </span>
-      {detailsPopover}
-    </>
-  );
-}
-
-function StoryWorkspace({
-  stories,
-  chapters,
-  lorebookEntries,
-  activeStoryId,
-  activeChapterId,
-  workspaceView,
-  chapterContent,
-  contextWindowInfo,
-  saveState,
-  generationStatus,
-  canvasStreaming,
-  smoothStreaming,
-  lorebookStatus,
-  writeReasoning,
-  lorebookReasoning,
-  lorebookThinking,
-  writeEditPreview,
-  canvasScrollPosition,
-  onOpenRail,
-  onOpenLorebook,
-  onBackToChapter,
-  onCanvasScrollPositionChange,
-  onChangeContent,
-  onCanvasImportFallback,
-  onCreateLorebookEntry,
-  onUpdateLorebookEntry,
-  onDeleteLorebookEntry,
-  onConfirmDeleteLorebookEntry,
-  onRepairTimeline,
-  onRepairLorebook,
-  onGenerateLorebookEntry,
-}) {
-  const canvasScrollRef = useRef(null);
-  const generationActiveRef = useRef(false);
-  const activeStory = stories.find((story) => story.id === activeStoryId);
-  const activeChapter = chapters.find((chapter) => chapter.id === activeChapterId);
-  const writingLocked = Boolean(generationStatus || lorebookStatus);
-
-  const writeStatus = generationStatus || lorebookStatus;
-  //the lorebook thinks in its own pass, so while it runs the dropdown shows its reasoning instead of the chapter's
-  const statusReasoning = lorebookThinking
-    ? lorebookReasoning
-    : generationStatus
-      ? writeReasoning
-      : null;
-  const statusEditPreview = lorebookThinking || !generationStatus ? null : writeEditPreview;
-  const {
-    markUserScroll: markCanvasScroll,
-    markWheelIntent: markCanvasWheelIntent,
-    markTouchStart: markCanvasTouchStart,
-    markTouchMove: markCanvasTouchMove,
-    scrollToBottom: scrollCanvasToBottom,
-    startFollowing: startCanvasFollowing,
-  } = useRafScroller(canvasScrollRef);
-
-  useLayoutEffect(() => {
-    if (workspaceView !== "chapter" || generationStatus) return undefined;
-
-    const canvas = canvasScrollRef.current;
-    if (!canvas) return undefined;
-
-    const maxScroll = Math.max(canvas.scrollHeight - canvas.clientHeight, 0);
-    const savedScrollTop = Number.isFinite(canvasScrollPosition) ? canvasScrollPosition : 0;
-    canvas.scrollTop = Math.min(Math.max(savedScrollTop, 0), maxScroll);
-
-    return undefined;
-    //generationStatus is read above but deliberately not a dependency: re-running the moment a run ends restores the offset we remembered before the last paragraph landed, and that backward jump reads as a scroll upward and kills auto follow
-  }, [activeChapterId, activeStoryId, canvasScrollPosition, workspaceView]);
-
-  const handleCanvasScroll = useCallback(() => {
-    markCanvasScroll();
-
-    const canvas = canvasScrollRef.current;
-    if (canvas) onCanvasScrollPositionChange(canvas.scrollTop);
-  }, [markCanvasScroll, onCanvasScrollPositionChange]);
-
-  useEffect(() => {
-    const generationActive = Boolean(generationStatus);
-    if (generationActive && !generationActiveRef.current) {
-      startCanvasFollowing();
-    }
-    generationActiveRef.current = generationActive;
-  }, [generationStatus, startCanvasFollowing]);
-
-  useEffect(() => {
-    if (!generationStatus) return;
-    scrollCanvasToBottom();
-  }, [chapterContent, generationStatus, scrollCanvasToBottom]);
-
-  if (!activeStory || !activeChapter) {
-    return (
-      <section className="min-h-0 overflow-y-auto px-4 py-20 sm:px-8 lg:px-10">
-        <div className="mx-auto flex min-h-[65dvh] max-w-4xl flex-col items-center justify-center text-center">
-          <button
-            type="button"
-            className="mb-6 inline-flex h-10 items-center justify-center rounded-full px-4 text-sm text-neutral-400 shadow-[var(--shadow-border)] hover:text-neutral-100 lg:hidden"
-            onClick={onOpenRail}
-          >
-            Open stories
-          </button>
-          <div className="text-2xl font-medium text-neutral-100">Choose a story and chapter</div>
-          <p className="mt-3 max-w-md text-sm leading-6 text-neutral-500">
-            Story mode keeps prose in the chapter canvas. The chat section is separate.
-          </p>
-        </div>
-      </section>
-    );
-  }
-
-  if (workspaceView === "lorebook" || workspaceView === "characters") {
-    return (
-      <StoryLorebook
-        story={activeStory}
-        chapters={chapters}
-        activeChapterId={activeChapterId}
-        entries={lorebookEntries}
-        onBack={onBackToChapter}
-        initialCategory={workspaceView === "characters" ? "character" : "all"}
-        onCreateEntry={onCreateLorebookEntry}
-        onUpdateEntry={onUpdateLorebookEntry}
-        onDeleteEntry={onDeleteLorebookEntry}
-        onConfirmDeleteEntry={onConfirmDeleteLorebookEntry}
-        onRepairTimeline={onRepairTimeline}
-        onRepairLorebook={onRepairLorebook}
-        onGenerateEntry={onGenerateLorebookEntry}
-        locked={writingLocked}
-      />
-    );
-  }
-
-  return (
-    <section
-      data-tour="write-chapter-canvas"
-      ref={canvasScrollRef}
-      onScroll={handleCanvasScroll}
-      onWheel={markCanvasWheelIntent}
-      onTouchStart={markCanvasTouchStart}
-      onTouchMove={markCanvasTouchMove}
-      className="write-canvas-scroll min-h-0 overflow-y-auto overscroll-contain px-4 pb-6 sm:px-8 lg:px-10"
-    >
-      <div className="mx-auto flex min-h-full w-full max-w-6xl flex-col">
-        <div className="flex min-h-0 flex-1 flex-col">
-          <div className="write-canvas-header mb-4 space-y-3">
-            <div className="flex items-start justify-between gap-3">
-              <h1 className="m-0 block min-w-0 truncate text-left text-2xl font-semibold leading-none text-neutral-100" title={activeChapter.title}>
-                {activeChapter.title}
-              </h1>
-              <div className="flex shrink-0 items-center gap-3 text-xs text-neutral-500">
-                {writeStatus ? (
-                  <WriteOperationStatus
-                    status={writeStatus}
-                    reasoning={statusReasoning?.text}
-                    reasoningStreaming={statusReasoning?.streaming}
-                    reasoningDurationMs={statusReasoning?.durationMs}
-                    editPreview={statusEditPreview}
-                  />
-                ) : (
-                  <span>{saveState || `${activeChapter.word_count || 0} words`}</span>
-                )}
-              </div>
-            </div>
-          </div>
-
-          {canvasStreaming && smoothStreaming ? (
-            <ChapterStreamingCanvas markdown={chapterContent} />
-          ) : (
-            <Suspense fallback={<div className="chapter-editor-loading" aria-hidden="true" />}>
-              <ChapterCanvasEditor
-                key={activeChapter.id}
-                chapterId={activeChapter.id}
-                markdown={chapterContent}
-                readOnly={writingLocked}
-                placeholder="Start writing here, or prompt the model to begin."
-                onChange={onChangeContent}
-                onImportFallback={onCanvasImportFallback}
-              />
-            </Suspense>
-          )}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function StoryWorkspaceBackButton({ onBack }) {
-  return (
-    <button
-      type="button"
-      onClick={onBack}
-      className={cx(
-        "inline-flex h-8 items-center gap-2 rounded-full bg-white/[0.035] px-3 text-xs font-medium text-neutral-300 shadow-[var(--shadow-border)] hover:bg-white/[0.06] hover:text-neutral-100 hover:shadow-[var(--shadow-border-hover)] focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/35",
-        CONTROL_MOTION,
-      )}
-    >
-      <ArrowLeft size={14} />
-      Back to chapter
-    </button>
-  );
-}
-
-function ComposerMenuButton({ label, detail, active = false, dataTour, disabled = false, onClick }) {
-  return (
-    <button
-      type="button"
-      role="menuitem"
-      data-tour={dataTour}
-      disabled={disabled}
-      onClick={onClick}
-      className={cx(
-        "flex min-h-10 w-full items-center justify-between gap-4 rounded-xl px-3 py-2 text-left text-sm transition-[background-color,color,scale] duration-150 ease-out hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-white/30 active:scale-[0.96] disabled:cursor-default disabled:hover:bg-transparent disabled:active:scale-100",
-        active ? "text-neutral-100" : "text-neutral-300",
-      )}
-    >
-      <span>{label}</span>
-      <span className={cx("max-w-[132px] truncate text-xs", active ? "text-neutral-400" : "text-neutral-500")}>
-        {detail}
-      </span>
-    </button>
-  );
-}
-
-function Composer({
-  value,
-  setValue,
-  disabled,
-  isStreaming,
-  settings,
-  models,
-  contextWindowInfo,
-  modelLocked,
-  onSubmit,
-  onStop,
-  onOpenSettings,
-  onToggleThinking,
-  openingMessage,
-  variant = "default",
-  forceShowThinking = false,
-  showContextMeter = true,
-  writeGenerationMode = null,
-  onToggleWriteGenerationMode,
-  writeHistoryEntries = [],
-  writeHistoryTitle = "Chapter history",
-  onOpenLorebook,
-  onOpenBrainstorm,
-  onUpdateLorebook,
-  onSetLorebookAuto,
-  lorebookUpdating = false,
-  systemPrompt = "",
-  onSaveSystemPrompt,
-  tourUi = null,
-  attachments = [],
-  attachmentsUploading = false,
-  onAttachFiles,
-  onRemoveAttachment,
-  dragActive = false,
-  webSearchEnabled = false,
-  onToggleWebSearch = null,
-}) {
-  const canThink = supportsThinking(models, settings.model) || forceShowThinking;
-  const canAttach = Boolean(onAttachFiles);
-  const canSeeImages = supportsImageInput(models, settings.model);
-  const canWebSearch = Boolean(onToggleWebSearch) && !writeGenerationMode;
-  const reasoningRequired = requiresThinking(models, settings.model);
-  const thinkingEnabled = effectiveThinkingEnabled(
-    models, settings.model, settings.thinking_enabled,
-  );
-  const thinkingStateLabel = thinkingEnabled
-    ? reasoningEffortLabel(models, settings.model, settings.reasoning_effort)
-    : "Instant";
-  const textareaRef = useRef(null);
-  const composerControlsRef = useRef(null);
-  const tourUiRef = useRef(null);
-  const [historyOpen, setHistoryOpen] = useState(false);
-  const [systemPromptOpen, setSystemPromptOpen] = useState(false);
-  const [contextMenuOpen, setContextMenuOpen] = useState(false);
-  const [modelMenuOpen, setModelMenuOpen] = useState(false);
-  const isEmptyVariant = variant === "empty";
-
-  useEffect(() => {
-    const textarea = textareaRef.current;
-    if (!textarea) return;
-    textarea.style.height = "auto";
-    const maxHeight = isEmptyVariant ? 184 : 126;
-    const nextHeight = Math.min(textarea.scrollHeight, maxHeight);
-    textarea.style.height = `${nextHeight}px`;
-    textarea.style.overflowY = textarea.scrollHeight > maxHeight ? "auto" : "hidden";
-  }, [isEmptyVariant, value]);
-
-  useEffect(() => {
-    function closeComposerMenus(event) {
-      if (event.key === "Escape") {
-        setContextMenuOpen(false);
-        setModelMenuOpen(false);
-      }
-    }
-
-    function closeOnOutsidePress(event) {
-      if (!composerControlsRef.current?.contains(event.target)) {
-        setContextMenuOpen(false);
-        setModelMenuOpen(false);
-      }
-    }
-
-    document.addEventListener("keydown", closeComposerMenus);
-    document.addEventListener("pointerdown", closeOnOutsidePress);
-    return () => {
-      document.removeEventListener("keydown", closeComposerMenus);
-      document.removeEventListener("pointerdown", closeOnOutsidePress);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!tourUi) {
-      if (tourUiRef.current) {
-        setContextMenuOpen(false);
-        setModelMenuOpen(false);
-        setSystemPromptOpen(false);
-        setHistoryOpen(false);
-      }
-      tourUiRef.current = null;
-      return;
-    }
-
-    tourUiRef.current = tourUi;
-    //every one of these lives inside the writing tools menu, so the menu has to be open for the tour to point at them
-    setContextMenuOpen(
-      ["tools", "generationMode", "lorebookMode", "lorebookUpdate"].includes(tourUi),
-    );
-    setModelMenuOpen(tourUi === "model");
-    setSystemPromptOpen(tourUi === "systemPrompt");
-    setHistoryOpen(tourUi === "history");
-  }, [tourUi]);
-
-  return (
-    <form
-      onSubmit={(event) => {
-        event.preventDefault();
-        isStreaming ? onStop() : onSubmit();
-      }}
-      className={cx(
-        isEmptyVariant
-          ? "pointer-events-none absolute inset-0 z-10 flex items-center justify-center px-4 pb-[12vh] pt-20 sm:px-8 lg:px-10"
-          : "bg-[#080808] px-4 py-4 sm:px-8 lg:px-10",
-        !isEmptyVariant && writeGenerationMode && "write-composer",
-      )}
-    >
-      <div className={cx("mx-auto w-full", isEmptyVariant ? "pointer-events-auto max-w-[760px]" : "max-w-4xl")}>
-        {isEmptyVariant && openingMessage && (
-          <div className="mb-8 text-center text-[22px] font-medium leading-tight text-neutral-200 sm:text-3xl">
-            {openingMessage}
-          </div>
-        )}
-        <div
-          className={cx(
-            "relative bg-[#141414]",
-            isEmptyVariant
-              ? "rounded-[30px]"
-              : "rounded-[24px]",
-          )}
-        >
-          <div className={cx(isEmptyVariant ? "px-[18px] pt-[15px] sm:px-[21px] sm:pt-[18px]" : "px-4 pt-3")}>
-            <textarea
-              ref={textareaRef}
-              value={value}
-              rows={1}
-              data-1p-ignore="true"
-              onChange={(event) => setValue(event.target.value)}
-              onKeyDown={(event) => {
-                if (event.key === "Enter" && !event.shiftKey) {
-                  event.preventDefault();
-                  isStreaming ? onStop() : onSubmit();
-                }
-              }}
-              placeholder={
-                isEmptyVariant
-                  ? "Ask anything"
-                  : writeGenerationMode
-                    ? `Ask ${promptModelName(models, settings.model)} to write anything`
-                    : `Ask ${promptModelName(models, settings.model)} anything`
-              }
-              className={cx(
-                "block w-full resize-none bg-transparent text-neutral-100 outline-none",
-                isEmptyVariant
-                  ? "max-h-[184px] min-h-[72px] text-base leading-7 placeholder:text-neutral-500 sm:text-lg sm:leading-8"
-                  : "max-h-[126px] min-h-6 text-sm leading-6 placeholder:text-neutral-600",
-              )}
-            />
-            {canAttach && (
-              <AttachmentChips
-                attachments={attachments}
-                uploading={attachmentsUploading}
-                onRemove={onRemoveAttachment}
-                compact={!isEmptyVariant}
-                className={isEmptyVariant ? "mt-3" : "mt-2.5"}
-              />
-            )}
-          </div>
-          <div
-            ref={composerControlsRef}
-            className={cx(
-              "flex items-center gap-2",
-              isEmptyVariant
-                ? "flex-wrap justify-between px-3 pb-[4.5px] pt-[3px] sm:flex-nowrap sm:px-[15px]"
-                : "justify-between px-4 pb-[5px] pt-[3px]",
-            )}
-            >
-            <div className="flex min-w-0 items-center gap-1.5">
-              <div className="flex items-center gap-0.5">
-              {canAttach && (
-                <AttachButton
-                  onFilesPicked={onAttachFiles}
-                  allowImages={canSeeImages}
-                  disabled={disabled || attachments.length >= MAX_FILES_PER_MESSAGE}
-                  modelName={promptModelName(models, settings.model)}
-                />
-              )}
-              {canWebSearch && (
-                <button
-                  type="button"
-                  onClick={onToggleWebSearch}
-                  aria-pressed={webSearchEnabled}
-                  aria-label="Web search"
-                  title={
-                    webSearchEnabled
-                      ? "Web search is on, OpenRouter bills each search"
-                      : "Search the web before answering"
-                  }
-                  className={cx(
-                    "grid h-8 w-8 shrink-0 place-items-center rounded-full hover:bg-white/[0.08] focus:outline-none",
-                    webSearchEnabled
-                      ? "text-blue-400"
-                      : "text-neutral-400",
-                    PROMPT_BAR_CONTROL_MOTION,
-                  )}
-                >
-                  <MaskIcon src="/icons/web.png" size={13} />
-                </button>
-              )}
-              </div>
-              {writeGenerationMode && (
-                <div className="relative">
-                  <button
-                    type="button"
-                    data-tour="write-tools-button"
-                    onClick={() => {
-                      setContextMenuOpen((open) => !open);
-                      setModelMenuOpen(false);
-                    }}
-                    aria-expanded={contextMenuOpen}
-                    aria-haspopup="menu"
-                    className={cx(
-                      "inline-flex h-8 min-w-0 items-center justify-center gap-1.5 rounded-full bg-transparent px-3 text-xs font-medium text-neutral-300 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-white/30 active:scale-[0.96]",
-                      PROMPT_BAR_CONTROL_MOTION,
-                    )}
-                  >
-                    <span>Writing tools</span>
-                    <span className="text-neutral-500">{WRITE_GENERATION_MODES[writeGenerationMode]}</span>
-                    <ChevronDown size={14} className={cx("writing-tools-chevron transition-transform duration-200", contextMenuOpen && "rotate-180")} />
-                  </button>
-                  {contextMenuOpen && (
-                    <div role="menu" className="absolute bottom-[calc(100%+8px)] -left-4 z-30 w-72 rounded-2xl bg-[#292929] p-1.5">
-                      <ComposerMenuButton label="Lorebook" detail="Story knowledge" onClick={() => { onOpenLorebook(); setContextMenuOpen(false); }} />
-                      <ComposerMenuButton label="Brainstorm" detail="Branch story ideas" onClick={() => { onOpenBrainstorm(); setContextMenuOpen(false); }} />
-                      <ComposerMenuButton label="System Prompt" detail={systemPrompt.trim() ? "Custom instructions" : "Default instructions"} onClick={() => { setSystemPromptOpen(true); setContextMenuOpen(false); }} active={Boolean(systemPrompt.trim())} />
-                      <ComposerMenuButton label="History" detail={`${writeHistoryEntries.length} saved events`} onClick={() => { setHistoryOpen(true); setContextMenuOpen(false); }} />
-                      <div className="my-1 border-t border-white/[0.08]" />
-                      <ComposerMenuButton dataTour="write-generation-mode" label={WRITE_GENERATION_MODES[writeGenerationMode]} detail="Switch writing action" onClick={onToggleWriteGenerationMode} />
-                      <ComposerMenuButton
-                        dataTour="write-lorebook-mode"
-                        label={LOREBOOK_UPDATE_MODES[settings.lorebook_auto ? "auto" : "manual"]}
-                        detail="Switch update mode"
-                        disabled={isStreaming || lorebookUpdating}
-                        onClick={() => onSetLorebookAuto(!settings.lorebook_auto)}
-                      />
-                      <ComposerMenuButton
-                        dataTour="write-lorebook-update"
-                        label="Update Lorebook"
-                        detail="Save story details"
-                        disabled={lorebookUpdating || isStreaming}
-                        onClick={() => { onUpdateLorebook(); setContextMenuOpen(false); }}
-                      />
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-            <div className="ml-auto flex min-w-0 items-center gap-1.5">
-              <div className="flex min-w-0 items-center gap-0">
-                {showContextMeter && <span data-tour="write-context-meter"><ContextWindowMeter info={contextWindowInfo} /></span>}
-                <div className="relative min-w-0">
-                <button
-                  type="button"
-                  data-tour="model-button"
-                  onClick={() => { setModelMenuOpen((open) => !open); setContextMenuOpen(false); }}
-                  aria-expanded={modelMenuOpen}
-                  aria-haspopup="menu"
-                  className={cx(
-                    "inline-flex min-w-0 max-w-[220px] items-center justify-center gap-1.5 rounded-full text-xs font-medium text-neutral-300 hover:bg-white/[0.08] focus:outline-none focus-visible:ring-2 focus-visible:ring-white/30 active:scale-[0.96] sm:max-w-[280px]",
-                    "h-8",
-                    "px-3 bg-transparent shadow-none",
-                    PROMPT_BAR_CONTROL_MOTION,
-                  )}
-                >
-                  <span className="truncate">{promptModelName(models, settings.model)}</span>
-                  {canThink && (
-                    <span className="hidden text-neutral-500 sm:inline">
-                      <span>{thinkingStateLabel}</span>
-                    </span>
-                  )}
-                  <ChevronDown size={14} className={cx("thinking-toggle-chevron shrink-0 transition-transform duration-200", modelMenuOpen && "rotate-180")} />
-                </button>
-                {modelMenuOpen && (
-                  <div role="menu" className="absolute bottom-[calc(100%+8px)] right-0 z-30 w-64 rounded-2xl bg-[#292929] p-1.5">
-                    <ComposerMenuButton
-                      label="Settings"
-                      detail={(
-                        <>
-                          {promptModelName(models, settings.model)}
-                          {modelLocked && <span className="ml-1 text-neutral-400">locked</span>}
-                        </>
-                      )}
-                      onClick={() => { onOpenSettings(); setModelMenuOpen(false); }}
-                    />
-                    {canThink && (
-                      <ComposerMenuButton
-                        label="Thinking"
-                        detail={(
-                          <>
-                            <span>{reasoningRequired ? "Required" : thinkingEnabled ? "On" : "Off"}</span>
-                          </>
-                        )}
-                        active={thinkingEnabled}
-                        dataTour="thinking-button"
-                        disabled={reasoningRequired}
-                        onClick={() => { onToggleThinking(); setModelMenuOpen(false); }}
-                      />
-                    )}
-                  </div>
-                )}
-                </div>
-              </div>
-
-            <button
-              type="submit"
-              data-tour="send-button"
-              disabled={!isStreaming && ((!value.trim() && attachments.length === 0) || disabled)}
-              className={cx(
-                "group relative inline-flex shrink-0 items-center justify-center rounded-full focus:outline-none focus-visible:ring-2 focus-visible:ring-accent/50",
-                "h-8 w-8 hover:bg-white/[0.08] disabled:hover:bg-transparent",
-                PROMPT_BAR_CONTROL_MOTION,
-                isStreaming
-                  ? "text-neutral-200"
-                  : cx(
-                    "text-neutral-300",
-                    "disabled:cursor-not-allowed disabled:text-neutral-600 disabled:active:scale-100",
-                  ),
-              )}
-              aria-label={isStreaming ? "Stop" : "Send"}
-              title={isStreaming ? "Stop" : "Send"}
-            >
-              <span
-                aria-hidden="true"
-                className={cx(
-                  "absolute inset-[3px] grid place-items-center overflow-hidden rounded-full transition-[opacity,filter,scale] duration-300 ease-[cubic-bezier(0.2,0,0,1)]",
-                  isStreaming
-                    ? "scale-100 opacity-100 blur-0"
-                    : "scale-[0.25] opacity-0 blur-[4px]",
-                )}
-              >
-                <Square size={13} />
-              </span>
-              <span
-                aria-hidden="true"
-                className={cx(
-                  "absolute inset-[3px] grid place-items-center overflow-hidden rounded-full transition-[opacity,filter,scale] duration-300 ease-[cubic-bezier(0.2,0,0,1)]",
-                  isStreaming
-                    ? "scale-[0.25] opacity-0 blur-[4px]"
-                    : "scale-100 opacity-100 blur-0",
-                )}
-              >
-                <i className="fi fi-rr-arrow-small-up send-arrow-icon" />
-              </span>
-            </button>
-            </div>
-          </div>
-          {canAttach && dragActive && (
-            <div
-              className={cx(
-                "attachment-drop-overlay pointer-events-none absolute inset-0 z-20 grid place-items-center",
-                "border border-dashed border-white/25 bg-[#141414]/95",
-                isEmptyVariant ? "rounded-[30px]" : "rounded-[24px]",
-              )}
-            >
-              <span className="flex items-center gap-2 text-sm font-medium text-neutral-300">
-                <Plus size={16} strokeWidth={2.25} />
-                {canSeeImages ? "Drop files to attach" : "Drop documents to attach"}
-              </span>
-            </div>
-          )}
-        </div>
-      </div>
-      <WriteHistoryModal
-        open={historyOpen}
-        entries={writeHistoryEntries}
-        title={writeHistoryTitle}
-        onClose={() => setHistoryOpen(false)}
-      />
-      <SystemPromptModal
-        open={systemPromptOpen}
-        value={systemPrompt}
-        onSave={onSaveSystemPrompt}
-        onClose={() => setSystemPromptOpen(false)}
-      />
-    </form>
-  );
-}
-
-function WriteHistoryModal({ open, entries, title, onClose }) {
-  const [rendered, setRendered] = useState(open);
-  const [phase, setPhase] = useState(open ? "open" : "closed");
-  const [expandedEntries, setExpandedEntries] = useState({});
-  const [openRuns, setOpenRuns] = useState({});
-  const closeRef = useRef(null);
-
-  useEffect(() => {
-    if (open) {
-      setRendered(true);
-      setPhase("open");
-      requestAnimationFrame(() => closeRef.current?.focus());
-      return undefined;
-    }
-
-    if (!rendered) return undefined;
-
-    setPhase("closing");
-    const closeMs =
-      parseFloat(
-        getComputedStyle(document.documentElement).getPropertyValue("--modal-close-dur"),
-      ) || 150;
-    const timeoutId = window.setTimeout(() => {
-      setRendered(false);
-      setPhase("closed");
-    }, closeMs);
-    return () => window.clearTimeout(timeoutId);
-  }, [open, rendered]);
-
-  useEffect(() => {
-    if (!rendered) return undefined;
-
-    function handleKeyDown(event) {
-      if (event.key === "Escape") onClose();
-    }
-
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [onClose, rendered]);
-
-  if (!rendered) return null;
-
-  const isOpen = phase === "open";
-  const runGroups = historyRunGroups(entries);
-  const eventCount = entries.filter((entry) => !isPromptEntry(entry)).length;
-  const totalCost = historyCostTotal(entries);
-  const totalWords = historyWordTotals(entries);
-  const lastRunId = runGroups.length > 0 ? runGroups[runGroups.length - 1].id : null;
-
-  function toggleExpanded(entryId) {
-    setExpandedEntries((current) => ({
-      ...current,
-      [entryId]: !current[entryId],
-    }));
-  }
-
-  //the newest run is open by default, so the flip has to resolve that default first or the
-  //first click on an untouched run just rewrites the state it already had
-  function toggleRun(runId) {
-    setOpenRuns((current) => ({
-      ...current,
-      [runId]: !(current[runId] ?? runId === lastRunId),
-    }));
-  }
-
-  return createPortal(
-    <div className="fixed inset-0 z-[80] grid place-items-center px-4 py-6">
-      <button
-        type="button"
-        aria-label="Close history"
-        className={cx(
-          "absolute inset-0 bg-black/60 backdrop-blur-sm transition-[opacity,backdrop-filter] duration-150 ease-out",
-          isOpen ? "opacity-100" : "opacity-0",
-        )}
-        onClick={onClose}
-      />
-      <section
-        data-tour="write-history"
-        role="dialog"
-        aria-modal="true"
-        aria-labelledby="write-history-title"
-        className={cx(
-          "t-modal relative z-10 flex max-h-[min(720px,calc(100dvh-2rem))] w-full max-w-[720px] flex-col overflow-hidden rounded-[26px] bg-[#191919] text-neutral-100 [box-shadow:var(--shadow-surface)]",
-          isOpen ? "is-open" : "is-closing",
-        )}
-      >
-        <header className="shrink-0 px-5 pb-4 pt-5 sm:px-6 sm:pt-6">
-          <div className="flex items-start justify-between gap-5">
-            <h2
-              id="write-history-title"
-              className="min-w-0 text-balance text-lg font-semibold leading-6 tracking-[-0.01em] text-neutral-100"
-            >
-              {title}
-            </h2>
-            <button
-              ref={closeRef}
-              type="button"
-              onClick={onClose}
-              className={cx(
-                "grid h-8 w-8 shrink-0 place-items-center rounded-full bg-transparent text-neutral-400 hover:bg-white/[0.08] hover:text-white focus:outline-none focus-visible:ring-2 focus-visible:ring-white/20",
-                PROMPT_BAR_CONTROL_MOTION,
-              )}
-              aria-label="Close history"
-            >
-              <X size={17} />
-            </button>
-          </div>
-
-          {eventCount === 0 && (
-            <p className="mt-1 text-sm leading-5 text-neutral-500">
-              Prompts and changes will appear here
-            </p>
-          )}
-        </header>
-
-        {/*this has to be the flex child itself, a percentage height inside one resolves against an
-           indefinite height and silently grows to the content instead of scrolling*/}
-        <div className="min-h-0 flex-1 overflow-y-auto px-3 pb-3 sm:px-4 sm:pb-4">
-          {runGroups.length > 0 ? (
-            <div className="space-y-2.5">
-              {runGroups.map((run, index) => (
-                <WriteHistoryRunAccordion
-                  key={run.id}
-                  run={run}
-                  index={index}
-                  open={openRuns[run.id] ?? run.id === lastRunId}
-                  onToggle={() => toggleRun(run.id)}
-                  expandedEntries={expandedEntries}
-                  onToggleEntry={toggleExpanded}
-                />
-              ))}
-            </div>
-          ) : (
-            <div className="rounded-[20px] bg-black/20 px-4 py-12 text-center shadow-[var(--shadow-border)]">
-              <div className="text-sm font-medium text-neutral-300">No history yet</div>
-              <div className="mt-1 text-xs leading-5 text-neutral-600">
-                Your next writing run will show up here.
-              </div>
-            </div>
-          )}
-        </div>
-
-        {eventCount > 0 && (
-          <footer className="shrink-0 border-t border-white/[0.06] px-5 py-3.5 sm:px-6">
-            {/*the dot between each total was 4px of glyph sitting between two 8px gaps, so dropping
-                it and widening the gap to 20px leaves the row spaced exactly where it was*/}
-            <div className="flex flex-wrap items-center gap-x-5 gap-y-1 text-[13px] leading-5 tabular-nums text-neutral-400">
-              <WriteHistoryTotal
-                value={formatInteger(runGroups.length)}
-                label={runGroups.length === 1 ? "prompt" : "prompts"}
-              />
-              <WriteHistoryTotal
-                value={formatInteger(eventCount)}
-                label={eventCount === 1 ? "action" : "actions"}
-              />
-              {(totalWords.added > 0 || totalWords.removed > 0) && (
-                <span className="flex items-center">
-                  {totalWords.added > 0 && (
-                    <span className="text-emerald-300">+{formatInteger(totalWords.added)}</span>
-                  )}
-                  {totalWords.added > 0 && totalWords.removed > 0 && (
-                    <span className="px-1 text-neutral-700">/</span>
-                  )}
-                  {totalWords.removed > 0 && (
-                    <span className="text-red-300">&minus;{formatInteger(totalWords.removed)}</span>
-                  )}
-                  <span className="ml-1.5 text-neutral-500">words</span>
-                </span>
-              )}
-              {totalCost > 0 && (
-                <span className="text-neutral-200">{formatCost(totalCost)}</span>
-              )}
-            </div>
-          </footer>
-        )}
-      </section>
-    </div>,
-    document.body,
-  );
-}
-
-function WriteHistoryTotal({ label, value }) {
-  return (
-    <span>
-      <span className="font-medium text-neutral-200">{value}</span>
-      <span className="ml-1.5 text-neutral-500">{label}</span>
-    </span>
-  );
-}
-
-function WriteHistoryRunAccordion({
-  run,
-  index,
-  open,
-  onToggle,
-  expandedEntries,
-  onToggleEntry,
-}) {
-  const actionCount = run.actions.length;
-  const activityRows = historyRunRows(run);
-  const runCost = historyCostTotal(run.actions);
-  const runWords = historyWordTotals(run.actions);
-
-  return (
-    <section
-      className="t-acc overflow-hidden rounded-[20px] bg-black/20 shadow-[var(--shadow-border)]"
-      data-open={String(open)}
-    >
-      {/*no CONTROL_MOTION here, the press scale it carries is a lot of movement for a row this wide
-          and the highlight it used to animate now lives on the title*/}
-      <button
-        type="button"
-        className="t-acc-head group flex w-full items-center justify-between gap-4 rounded-[20px] py-3 pl-3.5 pr-2.5 text-left focus:outline-none focus-visible:ring-2 focus-visible:ring-inset focus-visible:ring-white/20 sm:pl-4 sm:pr-3"
-        aria-expanded={open}
-        onClick={onToggle}
-      >
-        <span className="min-w-0">
-          <span className="-ml-1.5 inline-block rounded-lg px-1.5 text-sm font-semibold leading-5 text-neutral-100 transition-colors duration-150 ease-out group-hover:bg-white/[0.06]">
-            Prompt {index + 1}
-          </span>
-          <span className="mt-1 flex flex-wrap items-center gap-x-2.5 gap-y-1 text-[11px] font-medium leading-4 tabular-nums text-neutral-600">
-            <span>
-              {actionCount} {actionCount === 1 ? "action" : "actions"}
-            </span>
-            {runWords.added > 0 && <span className="text-emerald-300/80">+{formatInteger(runWords.added)}</span>}
-            {runWords.removed > 0 && <span className="text-red-300/80">&minus;{formatInteger(runWords.removed)}</span>}
-            {runCost > 0 && <span>{formatCost(runCost)}</span>}
-          </span>
-        </span>
-        <span className="t-acc-chevron grid h-9 w-9 shrink-0 place-items-center rounded-full text-neutral-500">
-          <ChevronDown size={17} strokeWidth={1.8} aria-hidden="true" />
-        </span>
-      </button>
-
-      <div className="t-acc-panel min-h-0">
-        <div className="t-acc-panel-inner min-h-0 px-3.5 pb-4 sm:px-4">
-          {activityRows.length > 0 ? (
-            <WriteHistoryActivity
-              rows={activityRows}
-              expandedEntries={expandedEntries}
-              onToggleEntry={onToggleEntry}
-            />
-          ) : (
-            <div className="px-1 text-xs leading-5 text-neutral-600">
-              No actions recorded for this prompt.
-            </div>
-          )}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-const HISTORY_DEPTH_TEXT = [
-  "text-sm font-medium leading-5 text-neutral-300",
-  "text-[13px] font-medium leading-5 text-neutral-400",
-  "text-[13px] font-normal leading-5 text-neutral-500",
-];
-
-function WriteHistoryActivity({ rows, expandedEntries, onToggleEntry }) {
-  return (
-    <WriteHistoryRows
-      rows={rows}
-      guides={[]}
-      parentFinal
-      expandedEntries={expandedEntries}
-      onToggleEntry={onToggleEntry}
-    />
-  );
-}
-
-function WriteHistoryRows({ rows, guides, parentFinal, expandedEntries, onToggleEntry }) {
-  return (
-    <ol className={guides.length === 0 ? "px-1" : undefined}>
-      {rows.map((row, index) => {
-        const lastSibling = index === rows.length - 1;
-        const children = historyRowChildren(row);
-        const expandable = children.length > 0 || Boolean(historyRowText(row));
-        const open = Boolean(expandedEntries[row.id]);
-        const expanded = expandable && open;
-
-        return (
-          <WriteHistoryAction
-            key={row.id}
-            row={row}
-            guides={guides}
-            expandable={expandable}
-            open={open}
-            //only the bottom row of everything on screen drops its gap, or the tree loses its rhythm
-            //wherever a branch happens to end
-            final={parentFinal && lastSibling && !expanded}
-            onToggle={() => onToggleEntry(row.id)}
-          >
-            {children.length > 0 && (
-              <WriteHistoryRows
-                rows={children}
-                //a guide is only drawn where its level still has rows to come
-                guides={[...guides, !lastSibling]}
-                parentFinal={parentFinal && lastSibling}
-                expandedEntries={expandedEntries}
-                onToggleEntry={onToggleEntry}
-              />
-            )}
-          </WriteHistoryAction>
-        );
-      })}
-    </ol>
-  );
-}
-
-function WriteHistoryAction({
-  row,
-  guides,
-  expandable,
-  open,
-  final,
-  onToggle,
-  children,
-}) {
-  //a row is either a branch or a leaf with detail, and each reads the same open flag for its own thing
-  const expanded = expandable && open;
-  const entry = row.entries[0];
-  const words = historyRowWords(row);
-  const depth = guides.length;
-  const textPanel = historyRowText(row);
-
-  const failed = entry.kind === "write_failed";
-  const textClass = HISTORY_DEPTH_TEXT[Math.min(depth, HISTORY_DEPTH_TEXT.length - 1)];
-
-  return (
-    <li className="t-tree" data-open={String(expanded)}>
-      <div className="t-tree-row grid grid-cols-[minmax(0,1fr)_auto] items-start gap-x-4">
-        {/*the guides are flex siblings of the label rather than a background, so they stretch to
-           whatever height the row ends up at once its detail text wraps*/}
-        <div className="flex min-w-0">
-          {/*a guide means "these hang off the row above", which nothing at the top level does, so the
-             outermost column is dropped entirely and every level below costs one step*/}
-          {guides.slice(1).map((drawn, level) => (
-            <WriteHistoryGuide key={level} drawn={drawn} />
-          ))}
-          {depth > 0 && <WriteHistoryGuide drawn />}
-
-          <div className={cx("min-w-0 flex-1", final ? "pb-0" : "pb-2.5")}>
-            {expandable ? (
-              <div className={cx("flex min-w-0 items-center gap-1", textClass)}>
-                <button
-                  type="button"
-                  onClick={onToggle}
-                  aria-expanded={expanded}
-                  className={cx(
-                    "-ml-1.5 flex min-w-0 max-w-full items-center gap-1.5 rounded-lg px-1.5 text-left transition-colors duration-150 ease-out hover:bg-white/[0.05] hover:text-neutral-100 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/20",
-                    textClass,
-                  )}
-                >
-                  <span className="truncate">{historyRowLabel(row)}</span>
-                  <span className="t-tree-chevron shrink-0 text-neutral-600">
-                    <ChevronDown size={13} strokeWidth={2} aria-hidden="true" />
-                  </span>
-                </button>
-                {textPanel?.text && (
-                  <WriteHistoryCopyButton
-                    text={textPanel.text}
-                    copyName={textPanel.kind === "thinking" ? "thinking" : "prompt"}
-                  />
-                )}
-              </div>
-            ) : (
-              <div className={cx("text-pretty", textClass, failed && "text-amber-200")}>
-                {historyRowLabel(row)}
-              </div>
-            )}
-
-            {!expandable && <WriteHistoryDetail entry={entry} />}
-          </div>
-        </div>
-
-        <WriteHistoryStats
-          added={words.added}
-          removed={words.removed}
-          isHide={entry.kind === "lore_hide"}
-        />
-      </div>
-
-      {expandable && (
-        <div className="t-tree-panel">
-          <div className="t-tree-panel-inner">
-            {textPanel ? (
-              <WriteHistoryTextPanel
-                text={textPanel.text}
-                emptyLabel={
-                  textPanel.kind === "thinking" ? "No thinking saved" : "No prompt saved"
-                }
-              />
-            ) : (
-              children
-            )}
-          </div>
-        </div>
-      )}
-    </li>
-  );
-}
-
-function WriteHistoryGuide({ drawn }) {
-  return (
-    <span className="relative w-[18px] shrink-0" aria-hidden="true">
-      {drawn && <span className="absolute inset-y-0 left-[5px] w-[2px] bg-white/[0.09]" />}
-    </span>
-  );
-}
-
-function WriteHistoryCopyButton({ text, copyName }) {
-  const [copyState, setCopyState] = useState("idle");
-  const timeoutRef = useRef(null);
-  const copyingRef = useRef(false);
-
-  useEffect(
-    () => () => window.clearTimeout(timeoutRef.current),
-    [],
-  );
-
-  async function copyText() {
-    if (!text || copyingRef.current) return;
-
-    copyingRef.current = true;
-    window.clearTimeout(timeoutRef.current);
-    setCopyState("copying");
-
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopyState("copied");
-    } catch {
-      setCopyState("failed");
-    } finally {
-      copyingRef.current = false;
-    }
-
-    timeoutRef.current = window.setTimeout(() => setCopyState("idle"), 2400);
-  }
-
-  return (
-    <div className="inline-flex shrink-0 items-center">
-      <button
-        type="button"
-        onClick={copyText}
-        disabled={copyState === "copying"}
-        aria-label={`Copy ${copyName}`}
-        className="inline-flex h-5 w-5 items-center justify-center rounded-md text-neutral-500 transition-colors hover:bg-white/[0.07] hover:text-neutral-100 focus:outline-none focus-visible:ring-2 focus-visible:ring-white/20 disabled:cursor-wait"
-      >
-        {copyState === "copied" ? <Check size="1em" className="text-emerald-300" aria-hidden="true" /> : <Copy size="1em" aria-hidden="true" />}
-      </button>
-      <span role="status" className={copyState === "failed" ? "ml-1 text-xs text-amber-200" : "sr-only"}>
-        {copyState === "copied" && `${copyName === "thinking" ? "Thinking" : "Prompt"} copied to clipboard.`}
-        {copyState === "failed" && "Couldn’t copy. Try again."}
-      </span>
-    </div>
-  );
-}
-
-function WriteHistoryTextPanel({ text, emptyLabel }) {
-  return (
-    <div className="mb-2.5 flex">
-      <WriteHistoryGuide drawn />
-      <div className={cx(
-        "min-w-0 max-h-[320px] flex-1 overflow-y-auto whitespace-pre-wrap break-words text-pretty text-[13px] leading-5",
-        text ? "cursor-text select-text text-neutral-300" : "text-neutral-600",
-      )}>
-        {text || emptyLabel}
-      </div>
-    </div>
-  );
-}
-
-function WriteHistoryStats({ added, removed, isHide = false }) {
-  if (!added && !removed) return null;
-
-  return (
-    <span className="flex shrink-0 items-center gap-2 self-start text-[11px] font-medium leading-5 tabular-nums">
-      {added > 0 && !isHide && <span className="text-emerald-300">+{added}</span>}
-      {removed > 0 && (
-        <span className={isHide ? "text-neutral-500" : "text-red-300"}>&minus;{removed}</span>
-      )}
-    </span>
-  );
-}
-
-function WriteHistoryDetail({ entry }) {
-  if (!entry.detail) return null;
-
-  return (
-    <div className="text-pretty text-xs leading-5 text-neutral-500">{entry.detail}</div>
-  );
-}
-
-function NewStoryModal({ open, onClose, onCreate }) {
-  return (
-    <NamePromptModal
-      open={open}
-      onClose={onClose}
-      onCreate={onCreate}
-      heading="Name your new story"
-      description="You can rename your story from the sidebar at any time"
-      placeholder="Story name"
-      inputLabel="Story name"
-      submitLabel="Create story"
-      dialogLabel="Close new story dialog"
-    />
   );
 }
